requirements_personal.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.7
django-allauth==0.40.0
django-ckeditor==5.8.0
django-crispy-forms==1.8.0
django-js-asset==1.2.2
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social==0.1.dev3
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
